全文检索后端实现

+++ ly-common/src/main/java/com/zl52074/leyou/common/pojo/PageResult.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public class PageResult<T> {
     private Long total;// 总条数
-    private Long totalPage;// 总页数
+    private Integer totalPage;// 总页数
     private List<T> items;// 当前页数据
 
     public PageResult() {
@@ -20,7 +20,7 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public PageResult(Long total, Long totalPage, List<T> items) {
+    public PageResult(Long total, Integer totalPage, List<T> items) {
         this.total = total;
         this.totalPage = totalPage;
         this.items = items;
@@ -42,11 +42,11 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public Long getTotalPage() {
+    public Integer getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(Long totalPage) {
+    public void setTotalPage(Integer totalPage) {
         this.totalPage = totalPage;
     }
 }
